Updated solution to save flow as integer.

+++ T2/src/IMaximumFlow.java
@@ -10,7 +10,7 @@ public interface IMaximumFlow {
 	 */
 	public class Solution {
 		Map<Integer, Set<Integer>> partition;
-		double maximumFlow;
+		int maximumFlow;
 	}
 
 	/**
